+++ lib-videoplayer/src/main/java/com/jik/lib/videoplayer/component/player/PlayerLoadingWheel.kt
@@ -1,4 +1,4 @@
-package com.jik.lib.videoplayer.component.thumnail
+package com.jik.lib.videoplayer.component.player
 
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.CircularProgressIndicator
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import com.jik.lib.videoplayer.component.iconSize
 
 @Composable
-fun ThumbnailLoadingWheel(
+fun PlayerLoadingWheel(
     modifier: Modifier = Modifier,
 ) {
     CircularProgressIndicator(
